Added Main Files
Added the manifest.json for setting up the extension to be read by Firefox with all related and needed files
Also for the phishingSearch.js this is where we are going to update everything and run our extension in. If any other files are needed we must update the "js" in the manifest.json file first so that Firefox detects it.
Note: This is missing the "permission" section where we still need to process this section with a lot of website, until then this is the latest update to it